+++ src/main/java/kr/modernworld/modernworldv2/global/common/StringToSenderReceiverNoFieldConverter.java
@@ -1,4 +1,4 @@
-package kr.modernworld.modernworldv2.user.presentation.present.dto.req;
+package kr.modernworld.modernworldv2.global.common;
 
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
