+++ src/main/java/kr/modernworld/modernworldv2/user/domain/port/present/PresentQueryRepository.java
@@ -1,7 +1,7 @@
 package kr.modernworld.modernworldv2.user.domain.port.present;
 
 import java.util.List;
-import kr.modernworld.modernworldv2.user.presentation.present.dto.req.SenderReceiverNoField;
+import kr.modernworld.modernworldv2.global.common.SenderReceiverNoField;
 import kr.modernworld.modernworldv2.user.presentation.present.dto.res.GetPresentResponseDTO;
 
 public interface PresentQueryRepository {
