+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/repository/present/PresentQueryRepositoryImpl.java
@@ -8,9 +8,9 @@ import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
+import kr.modernworld.modernworldv2.global.common.SenderReceiverNoField;
 import kr.modernworld.modernworldv2.user.domain.port.present.PresentQueryRepository;
 import kr.modernworld.modernworldv2.user.infrastructure.persistence.entity.QUserJPAEntity;
-import kr.modernworld.modernworldv2.user.presentation.present.dto.req.SenderReceiverNoField;
 import kr.modernworld.modernworldv2.user.presentation.present.dto.res.GetPresentResponseDTO;
 import kr.modernworld.modernworldv2.user.presentation.present.dto.res.PresentItemDTO;
 import kr.modernworld.modernworldv2.user.presentation.present.dto.res.PresentUserDTO;
